Employee-Attrition-master/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
_py-xgboost-mutex=2.0=cpu_0
alabaster=0.7.12=py37_0
argh=0.26.2=py37_0
astroid=2.3.3=py37_0
atomicwrites=1.4.0=py_0
attrs=19.3.0=py_0
autopep8=1.4.3=py37_0
babel=2.8.0=py_0
backcall=0.2.0=py_0
bcrypt=3.1.7=py37he774522_1
blas=1.0=mkl
bleach=3.1.5=py_0
brotlipy=0.7.0=py37he774522_1000
ca-certificates=2020.6.24=0
certifi=2020.6.20=py37_0
cffi=1.14.0=py37h7a1dbc1_0
chardet=3.0.4=py37_1003
click=7.1.2=py_0
cloudpickle=1.5.0=py_0
colorama=0.4.3=py_0
cryptography=2.9.2=py37h7a1dbc1_0
decorator=4.4.2=py_0
defusedxml=0.6.0=py_0
diff-match-patch=20181111=py_0
docutils=0.16=py37_1
entrypoints=0.3=py37_0
flake8=3.7.9=py37_0
flask=1.1.2=py_0
future=0.18.2=py37_1
gunicorn=20.0.4=pypi_0
icc_rt=2019.0.0=h0cc432a_1
icu=58.2=ha925a31_3
idna=2.10=py_0
imagesize=1.2.0=py_0
importlib-metadata=1.7.0=py37_0
importlib_metadata=1.7.0=0
intel-openmp=2020.1=216
intervaltree=3.0.2=py_0
ipykernel=5.3.0=py37h5ca1d4c_0
ipython=7.16.1=py37h5ca1d4c_0
ipython_genutils=0.2.0=py37_0
isort=4.3.21=py37_0
itsdangerous=1.1.0=py37_0
jedi=0.15.2=py37_0
jinja2=2.11.2=py_0
joblib=0.16.0=py_0
jpeg=9b=hb83a4c4_2
json5=0.9.5=py_0
jsonschema=3.2.0=py37_0
jupyter_client=6.1.5=py_0
jupyter_core=4.6.3=py37_0
jupyterlab=2.1.5=py_0
jupyterlab_server=1.2.0=py_0
keyring=21.2.1=py37_0
lazy-object-proxy=1.5.0=py37he774522_0
libpng=1.6.37=h2a8f88b_0
libsodium=1.0.18=h62dcd97_0
libspatialindex=1.9.3=h33f27b4_0
libxgboost=0.90=1
lightgbm=2.3.0=py37ha925a31_0
m2w64-gcc-libgfortran=5.3.0=6
m2w64-gcc-libs=5.3.0=7
m2w64-gcc-libs-core=5.3.0=7
m2w64-gmp=6.1.0=2
m2w64-libwinpthread-git=5.0.0.4634.697f757=2
markupsafe=1.1.1=py37he774522_0
mccabe=0.6.1=py37_1
mistune=0.8.4=py37he774522_0
mkl=2020.1=216
mkl-service=2.3.0=py37hb782905_0
mkl_fft=1.1.0=py37h45dec08_0
mkl_random=1.1.1=py37h47e9c7a_0
msys2-conda-epoch=20160418=1
nbconvert=5.6.1=py37_0
nbformat=5.0.7=py_0
notebook=6.0.3=py37_0
numpy=1.18.5=py37h6530119_0
numpy-base=1.18.5=py37hc3f5095_0
numpydoc=1.1.0=py_0
openssl=1.1.1g=he774522_0
packaging=20.4=py_0
pandas=1.0.5=py37h47e9c7a_0
pandoc=2.9.2.1=0
pandocfilters=1.4.2=py37_1
paramiko=2.7.1=py_0
parso=0.5.2=py_0
pathtools=0.1.2=py_1
pexpect=4.8.0=py37_0
pickleshare=0.7.5=py37_0
pip=20.1.1=py37_1
pluggy=0.13.1=py37_0
prometheus_client=0.8.0=py_0
prompt-toolkit=3.0.5=py_0
psutil=5.7.0=py37he774522_0
py-xgboost=0.90=py37_1
pycodestyle=2.5.0=py37_0
pycparser=2.20=py_0
pydocstyle=5.0.2=py_0
pyflakes=2.1.1=py37_0
pygments=2.6.1=py_0
pylint=2.4.4=py37_0
pynacl=1.4.0=py37h62dcd97_0
pyopenssl=19.1.0=py37_0
pyparsing=2.4.7=py_0
pyqt=5.9.2=py37h6538335_2
pyrsistent=0.16.0=py37he774522_0
pysocks=1.7.1=py37_0
python=3.7.7=h81c818b_4
python-dateutil=2.8.1=py_0
python-jsonrpc-server=0.3.4=py_0
python-language-server=0.31.10=py37_0
pytz=2020.1=py_0
pywin32=227=py37he774522_1
pywin32-ctypes=0.2.0=py37_1000
pywinpty=0.5.7=py37_0
pyyaml=5.3.1=py37he774522_1
pyzmq=19.0.1=py37ha925a31_1
qdarkstyle=2.8.1=py_0
qt=5.9.7=vc14h73c81de_0
qtawesome=0.7.2=py_0
qtconsole=4.7.5=py_0
qtpy=1.9.0=py_0
requests=2.24.0=py_0
rope=0.17.0=py_0
rtree=0.9.4=py37h21ff451_1
scikit-learn=0.23.1=py37h25d0782_0
scipy=1.5.0=py37h9439919_0
send2trash=1.5.0=py37_0
setuptools=47.3.1=py37_0
sip=4.19.8=py37h6538335_0
six=1.15.0=py_0
snowballstemmer=2.0.0=py_0
sortedcontainers=2.2.2=py_0
sphinx=3.1.2=py_0
sphinxcontrib-applehelp=1.0.2=py_0
sphinxcontrib-devhelp=1.0.2=py_0
sphinxcontrib-htmlhelp=1.0.3=py_0
sphinxcontrib-jsmath=1.0.1=py_0
sphinxcontrib-qthelp=1.0.3=py_0
sphinxcontrib-serializinghtml=1.1.4=py_0
spyder=4.1.3=py37_0
spyder-kernels=1.9.1=py37_0
sqlite=3.32.3=h2a8f88b_0
terminado=0.8.3=py37_0
testpath=0.4.4=py_0
threadpoolctl=2.1.0=pyh5ca1d4c_0
tornado=6.0.4=py37he774522_1
traitlets=4.3.3=py37_0
ujson=1.35=py37hfa6e2cd_0
urllib3=1.25.9=py_0
vc=14.1=h0510ff6_4
vs2015_runtime=14.16.27012=hf0eaf9b_3
watchdog=0.10.3=py37_0
wcwidth=0.2.5=py_0
webencodings=0.5.1=py37_1
werkzeug=1.0.1=py_0
wheel=0.34.2=py37_0
win_inet_pton=1.1.0=py37_0
wincertstore=0.2=py37_0
winpty=0.4.3=4
wrapt=1.12.1=py37he774522_1
yaml=0.2.5=he774522_0
yapf=0.30.0=py_0
zeromq=4.3.2=ha925a31_2
zipp=3.1.0=py_0
zlib=1.2.11=h62dcd97_4
